comment: 필요한 주석 추가 및 코드 정리

+++ cart_app/cart.py
@@ -15,6 +15,7 @@ class Cart:
             cart = self.session[settings.CART_SESSION_ID] = {}
         self.cart = cart
 
+
     def __iter__(self): # 물건을 또 담을 수 있게 함
         product_ids = self.cart.keys()
         # get the product objects and add them to the cart
@@ -27,8 +28,10 @@ class Cart:
             item['total_price'] = item['price'] * item['quantity']
             yield item
 
+
     def __len__(self): # 물건의 양 세기
         return sum(item['quantity'] for item in self.cart.values())
+
 
     def add(self, product, quantity=1, override_quantity=False):
         product_id = str(product.id)
@@ -41,9 +44,11 @@ class Cart:
             self.cart[product_id]['quantity'] += quantity
         self.save()
 
+
+    # mark the session as "modified" to make sure it gets saved
     def save(self):
-        # mark the session as "modified" to make sure it gets saved
         self.session.modified = True
+
 
     def remove(self, product):
         product_id = str(product.id)
@@ -51,10 +56,13 @@ class Cart:
             del self.cart[product_id]
             self.save()
 
+
+    # 세션에서 카트 삭제
     def clear(self):
-        # remove cart from session
         del self.session[settings.CART_SESSION_ID]
         self.save()
 
+
+    # 카트에 담긴 상품 총액 계산
     def get_total_price(self):
-        return sum(Decimal(item['price']) * item['quantity'] for item in self.cart.values())
+        return int(sum((item['price']) * item['quantity'] for item in self.cart.values()))
