+++ shop_app/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 from django.urls import reverse
 
+# 상품 카테고리
 class Category(models.Model):
     name = models.CharField(max_length=200)
     slug = models.SlugField(max_length=200,
@@ -22,6 +23,7 @@ class Category(models.Model):
                        args=[self.slug])
 
 
+# 상품
 class Product(models.Model):
     category = models.ForeignKey(Category,
                                  related_name='products',
@@ -30,8 +32,7 @@ class Product(models.Model):
     slug = models.SlugField(max_length=200)
     image = models.ImageField(upload_to='products/%Y/%m/%d', blank=True)
     description = models.TextField(blank=True)
-    price = models.DecimalField(max_digits=10,
-                                decimal_places=2)
+    price = models.IntegerField()
     available = models.BooleanField(default=True)
     created = models.DateTimeField(auto_now_add=True)
     updated = models.DateTimeField(auto_now=True)
